feat: modular eBPF sensors, decomposed tracer, and supply chain examples
Refactor the monolithic BPF program into modular headers (network, DNS,
file, process, SNI, cgroup) and split the tracer into dedicated event
handlers for each subsystem. Add GitHub Actions and Shai-Hulud supply
chain attack simulation examples with OPA rego policies.
