+++ internal/handlers/tracer/ebpf_maps.go
@@ -1,0 +1,257 @@
+package tracer
+
+// ebpf_maps.go handles population and maintenance of eBPF maps that control
+// kernel-side filtering behavior. These maps are the shared data plane between
+// userspace (this tracer) and the BPF programs running in the kernel.
+//
+// Map categories:
+//   - Allow lists: IPs, IPv6 addresses, hostnames, DNS servers
+//   - Block lists: executable names, protected file paths
+//   - Control: self TGID (to skip self-generated events)
+//   - Preloaded state: existing TCP connections preserved across BPF attach
+
+import (
+	"bufio"
+	"encoding/binary"
+	"net"
+	"os"
+	"strconv"
+	"strings"
+
+	"github.com/cilium/ebpf"
+
+	"github.com/kondukto-io/kntrl/internal/core/domain"
+	ebpfman "github.com/kondukto-io/kntrl/pkg/ebpf"
+	"github.com/kondukto-io/kntrl/pkg/logger"
+)
+
+// updateAllowedIPMaps writes all configured allowed IPv4 addresses into the
+// BPF map so that the kernel-side egress filter can allow them without
+// bouncing to userspace for policy evaluation.
+func updateAllowedIPMaps(allowedIPMap *ebpf.Map, arg *domain.Data) error {
+	for _, ipstr := range arg.AllowedIPs {
+		ip := ipstr.To4()
+		if ip == nil || len(ip) < 4 {
+			continue
+		}
+		ipUint32 := binary.LittleEndian.Uint32(ip[:4])
+		if err := allowedIPMap.Put(ipUint32, uint32(1)); err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+// updateAllowedIPv6Maps writes all configured allowed IPv6 addresses into the
+// BPF map as 16-byte keys.
+func updateAllowedIPv6Maps(allowedIPv6Map *ebpf.Map, arg *domain.Data) error {
+	for _, ip := range arg.AllowedIPv6s {
+		if len(ip) != net.IPv6len {
+			continue
+		}
+		var key [16]byte
+		copy(key[:], ip)
+		if err := allowedIPv6Map.Put(key, uint32(1)); err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+// updateAllowedHostMap writes all allowed hostnames into the BPF map as
+// fixed-size 256-byte keys (NUL-padded).
+func updateAllowedHostMap(allowedHostMap *ebpf.Map, arg *domain.Data) error {
+	for _, host := range arg.AllowedHosts {
+		var key [256]byte
+		copy(key[:], host)
+		if err := allowedHostMap.Put(key, uint32(1)); err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+// populateDNSServerMap writes the list of allowed DNS server IPs into the BPF
+// map so that DNS traffic to these servers is not blocked.
+func populateDNSServerMap(ebpfClient *ebpfman.EBPF, cmddata *domain.Data) {
+	allowedDNSServersMap := ebpfClient.Collection.Maps[domain.EBPFCollectionMapAllowedDNSServers]
+	if allowedDNSServersMap == nil || cmddata.AllowedDNSServers == nil {
+		return
+	}
+	for _, ip := range cmddata.AllowedDNSServers {
+		ipv4 := ip.To4()
+		if ipv4 == nil {
+			continue
+		}
+		ipUint32 := binary.LittleEndian.Uint32(ipv4)
+		if err := allowedDNSServersMap.Put(ipUint32, uint32(1)); err != nil {
+			logger.Log.Warnf("failed to update allowed DNS server: %v", err)
+		}
+	}
+}
+
+// populateSelfTGIDMap writes kntrl's own PID into the BPF map so that
+// kernel-side programs can drop events generated by the tracer itself,
+// preventing monitoring feedback loops.
+func populateSelfTGIDMap(ebpfClient *ebpfman.EBPF) {
+	selfTGIDMap := ebpfClient.Collection.Maps[domain.EBPFCollectionMapSelfTGID]
+	if selfTGIDMap == nil {
+		return
+	}
+	if err := selfTGIDMap.Put(uint32(0), uint32(os.Getpid())); err != nil {
+		logger.Log.Warnf("failed to set self TGID: %v", err)
+	}
+}
+
+// populateBlockedExecMap writes blocked executable names (max 16 bytes each)
+// into the BPF map for kernel-level process execution blocking.
+func populateBlockedExecMap(ebpfClient *ebpfman.EBPF, cmddata *domain.Data) {
+	blockedExecMap := ebpfClient.Collection.Maps[domain.EBPFCollectionMapBlockedExec]
+	if blockedExecMap == nil || cmddata.BlockedExecutables == nil {
+		return
+	}
+	for _, exe := range cmddata.BlockedExecutables {
+		var key [16]byte
+		copy(key[:], exe)
+		if err := blockedExecMap.Put(key, uint32(1)); err != nil {
+			logger.Log.Warnf("failed to update blocked exec map: %v", err)
+		}
+	}
+}
+
+// populateProtectedPathsMap writes protected file paths (max 256 bytes each)
+// into the BPF map for kernel-level file write protection.
+func populateProtectedPathsMap(ebpfClient *ebpfman.EBPF, cmddata *domain.Data) {
+	protectedPathsMap := ebpfClient.Collection.Maps[domain.EBPFCollectionMapProtectedPaths]
+	if protectedPathsMap == nil || cmddata.ProtectedPaths == nil {
+		return
+	}
+	for _, path := range cmddata.ProtectedPaths {
+		var key [256]byte
+		copy(key[:], path)
+		if err := protectedPathsMap.Put(key, uint32(1)); err != nil {
+			logger.Log.Warnf("failed to update protected paths map: %v", err)
+		}
+	}
+}
+
+// preloadEstablishedConns reads /proc/net/tcp and /proc/net/tcp6 to discover
+// active ESTABLISHED connections and pre-populates the BPF allowed IP maps.
+// This ensures that existing sessions (e.g. SSH connections used to manage the
+// host) survive when the egress BPF filter is attached.
+func preloadEstablishedConns(allowedIPMap, allowedIPv6Map *ebpf.Map) {
+	v4 := preloadTCP4(allowedIPMap)
+	var v6 int
+	if allowedIPv6Map != nil {
+		v6 = preloadTCP6(allowedIPv6Map)
+	}
+	if v4+v6 > 0 {
+		logger.Log.Infof("preloaded %d established connections (IPv4=%d, IPv6=%d)", v4+v6, v4, v6)
+	}
+}
+
+// preloadTCP4 parses /proc/net/tcp for ESTABLISHED (state 01) connections and
+// adds their remote IPv4 addresses to the BPF allowed IP map.
+func preloadTCP4(m *ebpf.Map) int {
+	f, err := os.Open("/proc/net/tcp")
+	if err != nil {
+		logger.Log.Debugf("preloadTCP4: %v", err)
+		return 0
+	}
+	defer f.Close()
+
+	var count int
+	scanner := bufio.NewScanner(f)
+	for scanner.Scan() {
+		fields := strings.Fields(scanner.Text())
+		if len(fields) < 4 {
+			continue
+		}
+		// Field 3 is the connection state; "01" = TCP_ESTABLISHED.
+		if fields[3] != "01" {
+			continue
+		}
+		// Field 2 is rem_address in hex format "XXXXXXXX:PPPP".
+		parts := strings.SplitN(fields[2], ":", 2)
+		if len(parts) < 1 || len(parts[0]) != 8 {
+			continue
+		}
+		ipHex, err := strconv.ParseUint(parts[0], 16, 32)
+		if err != nil {
+			continue
+		}
+		ipUint32 := uint32(ipHex)
+		if ipUint32 == 0 { // Skip 0.0.0.0
+			continue
+		}
+		if err := m.Put(ipUint32, uint32(1)); err != nil {
+			logger.Log.Debugf("preloadTCP4: put failed: %v", err)
+		} else {
+			count++
+		}
+	}
+	return count
+}
+
+// preloadTCP6 parses /proc/net/tcp6 for ESTABLISHED (state 01) connections and
+// adds their remote IPv6 addresses to the BPF allowed IPv6 map.
+func preloadTCP6(m *ebpf.Map) int {
+	f, err := os.Open("/proc/net/tcp6")
+	if err != nil {
+		logger.Log.Debugf("preloadTCP6: %v", err)
+		return 0
+	}
+	defer f.Close()
+
+	var count int
+	scanner := bufio.NewScanner(f)
+	for scanner.Scan() {
+		fields := strings.Fields(scanner.Text())
+		if len(fields) < 4 {
+			continue
+		}
+		if fields[3] != "01" {
+			continue
+		}
+		parts := strings.SplitN(fields[2], ":", 2)
+		if len(parts) < 1 || len(parts[0]) != 32 {
+			continue
+		}
+		hexStr := parts[0]
+
+		// Parse 32-char hex into [16]byte.
+		// /proc/net/tcp6 stores each 4-byte group in host (little-endian) order,
+		// so we reverse each group to get network byte order for the BPF map key.
+		var addr [16]byte
+		for i := 0; i < 4; i++ {
+			group := hexStr[i*8 : i*8+8]
+			for j := 0; j < 4; j++ {
+				b, err := strconv.ParseUint(group[j*2:j*2+2], 16, 8)
+				if err != nil {
+					break
+				}
+				// Reverse within each 4-byte group.
+				addr[i*4+(3-j)] = byte(b)
+			}
+		}
+
+		// Skip all-zeros (::).
+		allZero := true
+		for _, b := range addr {
+			if b != 0 {
+				allZero = false
+				break
+			}
+		}
+		if allZero {
+			continue
+		}
+
+		if err := m.Put(addr, uint32(1)); err != nil {
+			logger.Log.Debugf("preloadTCP6: put failed: %v", err)
+		} else {
+			count++
+		}
+	}
+	return count
+}
